fix:  创建企业同步创建用户并且做绑定

+++ rbac/rbac-manager/src/main/java/com/github/kop/rbac/repo/CompanyRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface CompanyRepository {
 
-  int create(RbacCompany rbacCompany);
+  Long create(RbacCompany rbacCompany);
 
   int update(RbacCompany rbacCompany);
